Generated SSL Cert
This generates the server's private key as well as the server's cert that gets sent to the browser. It's an RSA key, so it's widely used, but it's also 4096 bits with SHA256 for the signing so quite secure. The `-subj` bit at the end allows you to skip all the user required prompts that usually are needed.

Next step will be to figure out where the move the key / cert to within the docker image.